requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-decouple
passlib[bcrypt]
python-jose
motor